move ForeignKey type (#75)
