+++ src/core/src/schema/app/relation.rs
@@ -1,5 +1,5 @@
 mod belongs_to;
-pub use belongs_to::{BelongsTo, ForeignKey, ForeignKeyField};
+pub use belongs_to::BelongsTo;
 
 mod has_many;
 pub use has_many::HasMany;
