+++ src/core/src/schema/app/relation/belongs_to.rs
@@ -17,20 +17,6 @@ pub struct BelongsTo {
     pub foreign_key: ForeignKey,
 }
 
-#[derive(Debug, PartialEq)]
-pub struct ForeignKey {
-    pub fields: Vec<ForeignKeyField>,
-}
-
-#[derive(Debug, PartialEq)]
-pub struct ForeignKeyField {
-    /// The field on the source model that is acting as the foreign key
-    pub source: FieldId,
-
-    /// The field on the target model that this FK field maps to.
-    pub target: FieldId,
-}
-
 impl BelongsTo {
     pub fn target<'a>(&self, schema: &'a Schema) -> &'a Model {
         schema.model(self.target)
@@ -40,25 +26,5 @@ impl BelongsTo {
 impl From<BelongsTo> for FieldTy {
     fn from(value: BelongsTo) -> Self {
         FieldTy::BelongsTo(value)
-    }
-}
-
-impl ForeignKey {
-    pub(crate) fn placeholder() -> ForeignKey {
-        ForeignKey { fields: vec![] }
-    }
-
-    pub(crate) fn is_placeholder(&self) -> bool {
-        self.fields.is_empty()
-    }
-}
-
-impl ForeignKeyField {
-    pub fn source<'a>(&self, schema: &'a Schema) -> &'a Field {
-        schema.field(self.source)
-    }
-
-    pub fn target<'a>(&self, schema: &'a Schema) -> &'a Field {
-        schema.field(self.target)
     }
 }
